fix: Move all JavaScript-related files into js folder within components folder [ch10878]
