+++ polymer/src/elements/ip-reporting/ip-reporting-indicator-details.html
@@ -15,7 +15,7 @@
 <link rel="import" href="../../redux/store.html">
 <link rel="import" href="../../endpoints.html">
 
-<link rel="import" href="ip-reporting-indicator-details-functions.html">
+<link rel="import" href="js/ip-reporting-indicator-details-functions.html">
 
 <dom-module id="ip-reporting-indicator-details">
   <template>
