+++ polymer/src/elements/ip-reporting/js/ip-reporting-indicator-details.js
